rename CapiConfig file
