+++ devnets/devnetsHandler.ts
@@ -2,7 +2,7 @@ import * as path from "../deps/std/path.ts"
 import { $ } from "../mod.ts"
 import * as f from "../server/factories.ts"
 import { PermanentMemo } from "../util/memo.ts"
-import { Config } from "./CapiConfig.ts"
+import { Config } from "./Config.ts"
 import { devUserPublicKeys } from "./dev_users.ts"
 import { proxyWebSocket } from "./proxyWebSocket.ts"
 import { Network, startNetwork } from "./startNetwork.ts"
